xmptag: cast argument to isdigit to int
... as that is the specification and fixes compilation on Cygwin:

gstxmptaag.c: In function 'read_one_tag':
gstxmptag.c:1015: error: array subscript has type 'char'

+++ gst-libs/gst/tag/gstxmptag.c
@@ -1012,7 +1012,7 @@ read_one_tag (GstTagList * list, const gchar * tag, XmpTag * xmptag,
           gint num_digits = 0;
 
           /* find the number of digits */
-          while (isdigit (usec_str[num_digits++]) && num_digits < 6);
+          while (isdigit ((gint) usec_str[num_digits++]) && num_digits < 6);
 
           if (num_digits > 0) {
             /* fill up to 6 digits with 0 */
